+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Initial Mental Combat Skill Levels (MCSL) support. [#166](https://github.com/dmdorman/hero6e-foundryvtt/issues/166)
 - Fixed issue with large worlds failing to load.
 - Minor bug fixes for attacks created with "add attack" instead of via HDC upload.
+- Knocked Out when 0 STUN.
 
 
 # Version 3.0.6